Don't report processing time per directory by default
Prometheus metrics compress pretty well when they don't change
over time as often. This metric is almost guaranteed to change
each time, and thus will compress poorly. It's not particularly
helpful outside of debugging this specific exporter, and thus
will now be disabled by default.

Could also be the primary cause of
https://github.com/2i2c-org/infrastructure/issues/2930

Bumps version to 2.0 as this is technically a breaking change.